refactor($mobile): moving user apis into core wrapped by mobile
core functions are now core methods with mobile wrapper
